Init + Extra
Initial commit, started with a simple plane and 3 balls with their respective material. Then added a script to the camera so it could be locked to the white ball + added free movement for the mouse to move the camera. After that added a ray cast from the ball to the camera point foward, but the ball is not doing the correct movement with adding force.