Implemented Replay System
Implemented the replay system, fully functional.
Refactored some code in the BallController for a better user experience, implemented some events in classes to prevent only using boolean variables.
Created a play state manager which will controll wether the balls are in movement or not, to be able to start the record or stop the recording.